前台：组件结构调整
