+++ workHomeWebApp/src/app/app-routing.module.ts
@@ -15,31 +15,27 @@ const routes: Routes = [
         }
       },
       {
-        path: 'student',
-        loadChildren: () => import('./pages/student/student.module').then(m => m.StudentModule),
-        data: {
-          title: '学生管理'
-        }
-      },
-      {
-        path: 'item',
-        loadChildren: () => import('./pages/item/item.module').then(m => m.ItemModule),
-        data: {
-          title: '实验项目管理'
-        }
-      },
-      {
         path: 'personalCenter',
         loadChildren: () => import('./pages/personal-center/personal-center.module').then(m => m.PersonalCenterModule),
         data: {
           title: '个人中心'
         }
+      },
+      {
+        path: 'teacher',
+        loadChildren: () => import('./pages/teacher/teacher.module').then(m => m.TeacherModule),
+        data: {
+          title: '教师模块'
+        }
+      },
+      {
+        path: 'student',
+        loadChildren: () => import('./pages/student/student-routing.module').then(m => m.StudentRoutingModule),
+        data: {
+          title: '学生模块'
+        }
       }
     ]
-  },
-  {
-    path: 'work',
-    loadChildren: () => import('./pages/work/work.module').then(m => m.WorkModule)
   }
 ];
