+++ workHomeWebApp/src/app/pages/student/work/work.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { IndexComponent } from './index/index.component';
 import { EditComponent } from './edit/edit.component';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import { FuncModule } from '../../func/func.module';
+import { FuncModule } from '../../../func/func.module';
 import { RouterModule } from '@angular/router';
 import { WorkRoutingModule } from './work-routing.module';
