+++ workHomeWebApp/src/app/pages/teacher/item/edit/edit.component.spec.ts
@@ -3,8 +3,8 @@ import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 import {EditComponent} from './edit.component';
 import {ReactiveFormsModule} from '@angular/forms';
 import {RouterTestingModule} from '@angular/router/testing';
-import {ServiceTestingModule} from '../../../service/service-tesing/service-testing.module';
-import {AppTestingModule} from '../../../app-testing/app-testing.module';
+import {ServiceTestingModule} from '../../../../service/service-tesing/service-testing.module';
+import {AppTestingModule} from '../../../../app-testing/app-testing.module';
 import {DatePipe} from '@angular/common';
 
 describe('EditComponent', () => {
