+++ workHomeWebApp/src/app/pages/teacher/item/edit/edit.component.ts
@@ -1,12 +1,12 @@
 import {Component, OnInit} from '@angular/core';
 import {AbstractControl, FormBuilder, FormGroup, Validators} from '@angular/forms';
-import {Student} from '../../../common/student';
+import {Student} from '../../../../common/student';
 import {ActivatedRoute, Params, Router} from '@angular/router';
-import {StudentService} from '../../../service/student.service';
-import {AppComponent} from '../../../app.component';
+import {StudentService} from '../../../../service/student.service';
+import {AppComponent} from '../../../../app.component';
 import {HttpErrorResponse} from '@angular/common/http';
-import {ItemService} from '../../../service/item.service';
-import {Item} from '../../../common/item';
+import {ItemService} from '../../../../service/item.service';
+import {Item} from '../../../../common/item';
 import {DatePipe} from '@angular/common';
 
 @Component({
