+++ workHomeWebApp/src/app/pages/teacher/student/student-edit/student-edit.component.html
