+++ workHomeWebApp/src/app/pages/teacher/student/student-edit/student-edit.component.spec.ts
@@ -1,22 +1,22 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { AddComponent } from './add.component';
+import { StudentEditComponent } from './student-edit.component';
 import {ReactiveFormsModule} from '@angular/forms';
-import {ServiceTestingModule} from '../../../service/service-tesing/service-testing.module';
 import {RouterTestingModule} from '@angular/router/testing';
-import {AppTestingModule} from '../../../app-testing/app-testing.module';
+import {ServiceTestingModule} from '../../../../service/service-tesing/service-testing.module';
+import {AppTestingModule} from '../../../../app-testing/app-testing.module';
 
-describe('AddComponent', () => {
-  let component: AddComponent;
-  let fixture: ComponentFixture<AddComponent>;
+describe('StudentEditComponent', () => {
+  let component: StudentEditComponent;
+  let fixture: ComponentFixture<StudentEditComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ AddComponent ],
+      declarations: [ StudentEditComponent ],
       imports: [
         ReactiveFormsModule,
-        ServiceTestingModule,
         RouterTestingModule,
+        ServiceTestingModule,
         AppTestingModule
       ]
     })
@@ -24,7 +24,7 @@ describe('AddComponent', () => {
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(AddComponent);
+    fixture = TestBed.createComponent(StudentEditComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
