+++ workHomeWebApp/src/app/pages/teacher/student/student-edit/student-edit.component.ts
@@ -1,10 +1,10 @@
 import {Component, OnInit} from '@angular/core';
 import {AbstractControl, FormBuilder, FormGroup, Validators} from '@angular/forms';
 import {ActivatedRoute, Params, Router} from '@angular/router';
-import {StudentService} from '../../../service/student.service';
-import {Student} from '../../../common/student';
+import {StudentService} from '../../../../service/student.service';
+import {Student} from '../../../../common/student';
 import {HttpErrorResponse} from '@angular/common/http';
-import {AppComponent} from '../../../app.component';
+import {AppComponent} from '../../../../app.component';
 
 @Component({
   selector: 'app-student-edit',
