+++ workHomeWebApp/src/app/service/menu.service.ts
@@ -8,9 +8,9 @@ import {BehaviorSubject, Observable} from 'rxjs';
 export class MenuService {
   private menusSubject = new BehaviorSubject<Array<Menu>>([
     new Menu({name: '首页', url: '', roles: [Menu.ROLE_TEACHER, Menu.ROLE_STUDENT]}),
-    new Menu({name: '我的作业', url: 'work', roles: [Menu.ROLE_STUDENT]}),
-    new Menu({name: '学生管理', url: 'student', roles: [Menu.ROLE_TEACHER]}),
-    new Menu({name: '实验项目管理', url: 'item', roles: [Menu.ROLE_TEACHER]}),
+    new Menu({name: '我的作业', url: 'student/work', roles: [Menu.ROLE_STUDENT]}),
+    new Menu({name: '学生管理', url: 'teacher/student', roles: [Menu.ROLE_TEACHER]}),
+    new Menu({name: '实验项目管理', url: 'teacher/item', roles: [Menu.ROLE_TEACHER]}),
     new Menu({name: '个人中心', url: 'personalCenter', roles: [Menu.ROLE_TEACHER, Menu.ROLE_STUDENT]})
   ]);
